Add files via upload
This project focuses on the social network relationships between users on meetup.com located in Nashville. Multiple algorithms and evaluation methods like graph-based social network analysis and centrality measures were exercised to fulfill the purpose of disclosing the reality in the social network media. Rely on the graph and diversified measures, the differing in the distribution of social network resources were demonstrated as a “social gap” between distinct kind of participators.